ft817: Move relevant code fomr header to C file. Dropped pointless NULL check before free. Explicitly implemented digi_mode instead of awkward add on to fm_status Handle unlikely case of getting bogus digi mode, this prevent a garbage read while determining the bandwidth of the mode
