+++ rigs/yaesu/ft817.h
@@ -76,82 +76,7 @@
 #define FT817_CACHE_TIMEOUT     50
 
 
-enum ft817_native_cmd_e
-{
-    FT817_NATIVE_CAT_LOCK_ON = 0,
-    FT817_NATIVE_CAT_LOCK_OFF,
-    FT817_NATIVE_CAT_PTT_ON,
-    FT817_NATIVE_CAT_PTT_OFF,
-    FT817_NATIVE_CAT_SET_FREQ,
-    FT817_NATIVE_CAT_SET_MODE_LSB,
-    FT817_NATIVE_CAT_SET_MODE_USB,
-    FT817_NATIVE_CAT_SET_MODE_CW,
-    FT817_NATIVE_CAT_SET_MODE_CWR,
-    FT817_NATIVE_CAT_SET_MODE_AM,
-    FT817_NATIVE_CAT_SET_MODE_FM,
-    FT817_NATIVE_CAT_SET_MODE_FM_N,
-    FT817_NATIVE_CAT_SET_MODE_DIG,
-    FT817_NATIVE_CAT_SET_MODE_PKT,
-    FT817_NATIVE_CAT_CLAR_ON,
-    FT817_NATIVE_CAT_CLAR_OFF,
-    FT817_NATIVE_CAT_SET_CLAR_FREQ,
-    FT817_NATIVE_CAT_SET_VFOAB,
-    FT817_NATIVE_CAT_SPLIT_ON,
-    FT817_NATIVE_CAT_SPLIT_OFF,
-    FT817_NATIVE_CAT_SET_RPT_SHIFT_MINUS,
-    FT817_NATIVE_CAT_SET_RPT_SHIFT_PLUS,
-    FT817_NATIVE_CAT_SET_RPT_SHIFT_SIMPLEX,
-    FT817_NATIVE_CAT_SET_RPT_OFFSET,
-    FT817_NATIVE_CAT_SET_DCS_ON,
-    FT817_NATIVE_CAT_SET_CTCSS_ON,
-    FT817_NATIVE_CAT_SET_CTCSS_ENC_ON,
-    FT817_NATIVE_CAT_SET_CTCSS_DCS_OFF,
-    FT817_NATIVE_CAT_SET_CTCSS_FREQ,
-    FT817_NATIVE_CAT_SET_DCS_CODE,
-    FT817_NATIVE_CAT_GET_RX_STATUS,
-    FT817_NATIVE_CAT_GET_TX_STATUS,
-    FT817_NATIVE_CAT_GET_FREQ_MODE_STATUS,
-    FT817_NATIVE_CAT_PWR_WAKE,
-    FT817_NATIVE_CAT_PWR_ON,
-    FT817_NATIVE_CAT_PWR_OFF,
-    FT817_NATIVE_CAT_EEPROM_READ,
-    FT817_NATIVE_SIZE       /* end marker */
-};
-
-
-typedef enum ft817_native_cmd_e ft817_native_cmd_t;
-
-
-/* fixme: why declare static? it has no effect */
-static int ft817_init(RIG *rig);
-static int ft817_open(RIG *rig);
-static int ft817_cleanup(RIG *rig);
-static int ft817_close(RIG *rig);
-static int ft817_set_freq(RIG *rig, vfo_t vfo, freq_t freq);
-static int ft817_get_freq(RIG *rig, vfo_t vfo, freq_t *freq);
-static int ft817_set_mode(RIG *rig, vfo_t vfo, rmode_t mode, pbwidth_t width);
-static int ft817_get_mode(RIG *rig, vfo_t vfo, rmode_t *mode, pbwidth_t *width);
-static int ft817_set_ptt(RIG *rig, vfo_t vfo, ptt_t ptt);
-static int ft817_get_ptt(RIG *rig, vfo_t vfo, ptt_t *ptt);
-static int ft817_get_level(RIG *rig, vfo_t vfo, setting_t level, value_t *val);
-static int ft817_set_func(RIG *rig, vfo_t vfo, setting_t func, int status);
-static int ft817_set_dcs_code(RIG *rig, vfo_t vfo, tone_t code);
-static int ft817_set_ctcss_tone(RIG *rig, vfo_t vfo, tone_t tone);
-static int ft817_set_dcs_sql(RIG *rig, vfo_t vfo, tone_t code);
-static int ft817_set_ctcss_sql(RIG *rig, vfo_t vfo, tone_t tone);
-static int ft817_set_rptr_shift(RIG *rig, vfo_t vfo, rptr_shift_t shift);
-static int ft817_set_rptr_offs(RIG *rig, vfo_t vfo, shortfreq_t offs);
-static int ft817_set_rit(RIG *rig, vfo_t vfo, shortfreq_t rit);
-static int ft817_get_dcd(RIG *rig, vfo_t vfo, dcd_t *dcd);
 int ft817_set_powerstat(RIG *rig, powerstat_t status);
-static int ft817_vfo_op(RIG *rig, vfo_t vfo, vfo_op_t op);
-static int ft817_get_split_vfo(RIG *rig, vfo_t vfo, split_t *split,
-                               vfo_t *tx_vfo);
-static int ft817_set_split_vfo(RIG *rig, vfo_t vfo, split_t split,
-                               vfo_t tx_vfo);
-static int ft817_power2mW(RIG *rig, unsigned int *mwpower, float power,
-                          freq_t freq, rmode_t mode);
-static int ft817_mW2power(RIG *rig, float *power, unsigned int mwpower,
-                          freq_t freq, rmode_t mode);
+int ft817_read_ack(RIG *rig);
 
 #endif /* _FT817_H */
